Move monkey patching into dedicated module
